requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
firebase_admin==7.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
